Fixed Literally Every Mess
I cannot even start to talk about the changes that have been made to this file, it was 4 hours of unpaid labor done by Ece and Myself. I send my deepest regards to the female family members of anyone who had a finger in the bugs that I had to fix through so much work.